+++ src/modules/tools/draw/store/actions/settersDraw.js
@@ -25,7 +25,7 @@ async function setActive ({state, commit, dispatch, rootState}, active) {
 
     if (active) {
         commit("setSymbol", state.iconList[0]);
-        commit("setLayer", await dispatch("Maps/addNewLayerIfNotExists", "import_draw_layer", {root: true}));
+        commit("setLayer", await dispatch("Maps/addNewLayerIfNotExists", "importDrawLayer", {root: true}));
         commit("setImgPath", rootState?.configJs?.wfsImgPath);
         dispatch("createDrawInteractionAndAddToMap", {active: state.currentInteraction === "draw"});
         dispatch("createSelectInteractionAndAddToMap", state.currentInteraction === "delete");