+++ src/modules/tools/draw/tests/unit/store/withoutGUIDraw.spec.js
@@ -158,7 +158,7 @@ describe("src/modules/tools/draw/store/actions/withoutGUIDraw.js", () => {
             expect(commit.secondCall.args).to.eql(["setWithoutGUI", true]);
             expect(commit.thirdCall.args).to.eql(["setDrawType", {id: "drawSymbol", geometry: "Point"}]);
             expect(commit.lastCall.args).to.eql(["setLayer", undefined]);
-            expect(dispatch.firstCall.args).to.eql(["Maps/addNewLayerIfNotExists", "import_draw_layer", {root: true}]);
+            expect(dispatch.firstCall.args).to.eql(["Maps/addNewLayerIfNotExists", "importDrawLayer", {root: true}]);
             expect(dispatch.secondCall.args).to.eql(["createDrawInteractionAndAddToMap", {active: true, maxFeatures}]);
             expect(dispatch.thirdCall.args).to.eql(["createSelectInteractionAndAddToMap", false]);
             expect(dispatch.getCall(3).args).to.eql(["createModifyInteractionAndAddToMap", false]);
@@ -175,7 +175,7 @@ describe("src/modules/tools/draw/store/actions/withoutGUIDraw.js", () => {
             expect(commit.thirdCall.args).to.eql(["setDrawType", {id: "drawLine", geometry: "LineString"}]);
             expect(commit.getCall(3).args).to.eql(["setLineStringSettings", {color, colorContour: color}]);
             expect(commit.lastCall.args).to.eql(["setLayer", undefined]);
-            expect(dispatch.firstCall.args).to.eql(["Maps/addNewLayerIfNotExists", "import_draw_layer", {root: true}]);
+            expect(dispatch.firstCall.args).to.eql(["Maps/addNewLayerIfNotExists", "importDrawLayer", {root: true}]);
             expect(dispatch.secondCall.args).to.eql(["createDrawInteractionAndAddToMap", {active: true, maxFeatures}]);
             expect(dispatch.thirdCall.args).to.eql(["createSelectInteractionAndAddToMap", false]);
             expect(dispatch.getCall(3).args).to.eql(["createModifyInteractionAndAddToMap", false]);
@@ -195,7 +195,7 @@ describe("src/modules/tools/draw/store/actions/withoutGUIDraw.js", () => {
             expect(commit.thirdCall.args).to.eql(["setDrawType", {id: "drawArea", geometry: "Polygon"}]);
             expect(commit.getCall(3).args).to.eql(["setPolygonSettings", {color: resultColor, opacity}]);
             expect(commit.lastCall.args).to.eql(["setLayer", undefined]);
-            expect(dispatch.firstCall.args).to.eql(["Maps/addNewLayerIfNotExists", "import_draw_layer", {root: true}]);
+            expect(dispatch.firstCall.args).to.eql(["Maps/addNewLayerIfNotExists", "importDrawLayer", {root: true}]);
             expect(dispatch.secondCall.args).to.eql(["createDrawInteractionAndAddToMap", {active: true, maxFeatures}]);
             expect(dispatch.thirdCall.args).to.eql(["createSelectInteractionAndAddToMap", false]);
             expect(dispatch.getCall(3).args).to.eql(["createModifyInteractionAndAddToMap", false]);