+++ src/modules/tools/measure/store/stateMeasure.js
@@ -70,8 +70,8 @@ const state = {
     layer: new VectorLayer({
         source,
         style,
-        id: "measure_layer",
-        name: "measure_layer",
+        id: "measureLayer",
+        name: "measureLayer",
         alwaysOnTop: true
     }),
     featureId: null,