+++ src/modules/tools/measure/tests/end2end/Measure.e2e.js
@@ -103,9 +103,9 @@ async function MeasureTests ({builder, url, resolution, mode, capability}) {
                     });
 
                     it("allows deleting made measurements by clicking the deletion button", async function () {
-                        expect(await driver.executeScript(hasVectorLayerLength, "measure_layer", 0)).to.be.false;
+                        expect(await driver.executeScript(hasVectorLayerLength, "measureLayer", 0)).to.be.false;
                         await deleteButton.click();
-                        expect(await driver.executeScript(hasVectorLayerLength, "measure_layer", 0)).to.be.true;
+                        expect(await driver.executeScript(hasVectorLayerLength, "measureLayer", 0)).to.be.true;
                     });
 
                     it("draws a polygon, ending in double-click, displaying area in meters² and deviation", async function () {