+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Changed
 - CompareFeatures: The values true and false are now translated.
 - The interactions between the `GFI` and the `Map` have been moved from the Map module to the GFI module/tool.
+- Rename the layer names `measure_layer` to `measureLayer` and draw `import_draw_layer` to `importDrawLayer`.
 
 ### Deprecated